Fix the RTC_T definition

+++ c/libcr2/syscall.h
@@ -61,12 +61,12 @@ typedef struct {
  *  This structure is to hold all important fields needed to read time from the RTC hardware chip.
  */
 typedef struct {
-	uint16_t seconds;
+	uint8_t seconds;
 	uint8_t minutes;
 	uint8_t hours;
 	uint8_t day;
 	uint8_t month;
-	uint8_t year;
+	uint16_t year;
 } __attribute__((packed)) RTC_T;
 
 /*
